+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file
 ## Unreleased
 
 - add TUFX compat
+- add ClickThroughBlocker compat
 
 
 ## 0.0.5 - 2025-07-28